gbp-repack-mozjs-source: Include ICU and changed files
Some files are not included in the debian/*/master tree but are still
kept in the orig file, so it makes sense to have them around, while
others (Cargo.toml) gets modified in place by the mozilla mozjs
generator script. Even if we want to use this version, unfortunately we
need to do this via distro-patching or we'll get an archives mismatch.