gbp.conf: Use native gbp postunpack repacking
This is supported as part of [1] and allows better repacking using gbp
without having to repeat the same operations twice, and without actually
changing the orig file.

We can move the filtering into the script, even if it may involve some
bigger temporary folder, it's still faster and doesn't require
duplicating places where we handle it.

[1] https://github.com/agx/git-buildpackage/pull/69