debian/gbp-repack-source: Also filter files excluded in debian/copyright
Those files are already removed in case uscan is used, but not in case
we import the original tar manually, so let's mimic that.

Even if it's not the tool made for parsing copyright files, grep-dctrl
works great to give us a list without having to handle the parsing
manually.